pyproject.toml
[project]
name = "adk-multiagent-mcp-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "geopy>=2.4.1",
    "google-adk>=0.1.0",
    "google-cloud>=0.34.0",
    "mcp>=1.6.0",
    "google-generativeai>=0.8.5",
    "uvicorn>=0.34.1",
    "tavily-python>=0.5.4",
    "google-cloud-speech>=2.32.0",
    "moviepy>=2.1.2",
    "pydub>=0.2.25.1",
    "pandas>=2.2.3",
]
